dev - Started development of the last version
---

This version, 0.0.4.10, is the last version of RetroKS that will be released this year. Nitrocid KS 0.1.0 Beta 3 is going to be the last version that's going to have RetroKS support.

Starting with the release of the RC version of N-KS 0.1.0, we'll move this repo to the archives and will not continue development. Also, we'll remove its associated addon.

---

Type: dev
Breaking: False
Doc Required: False
Part: 1/1